feat(plugin-workflow-loop): add more configuration
