+++ packages/plugins/@nocobase/plugin-workflow-loop/src/constants.ts
@@ -1,0 +1,24 @@
+/**
+ * This file is part of the NocoBase (R) project.
+ * Copyright (c) 2020-2024 NocoBase Co., Ltd.
+ * Authors: NocoBase Team.
+ *
+ * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
+ * For more information, please refer to: https://www.nocobase.com/agreement.
+ */
+
+export const CHECKPOINT = {
+  BEFORE: 0,
+  AFTER: 1,
+};
+
+export const CONTINUE_ON_FALSE = {
+  BREAK: false,
+  CONTINUE: true,
+};
+
+export const EXIT = {
+  CONTINUE: 2,
+  BREAK: 1,
+  RETURN: 0,
+};
