+++ packages/plugins/@nocobase/plugin-workflow-loop/src/locale/zh-CN.json
@@ -4,5 +4,23 @@
   "Loop index": "当前索引",
   "Loop length": "循环长度",
   "By using a loop node, you can perform the same operation on multiple sets of data. The source of these sets can be either multiple records from a query node or multiple associated records of a single record. Loop node can also be used for iterating a certain number of times or for looping through each character in a string. However, excessive looping may cause performance issues, so use with caution.": "使用循环节点可以对多条数据进行同样的操作，多条数据的来源可以是查询节点的多条结果，或者一条数据的多条关系数据。也可以用于一定次数的循环，或者对字符串中每一个字符的循环处理。循环次数过高可能引起性能问题，请谨慎使用。",
-  "A single number will be treated as a loop count, a single string will be treated as an array of characters, and other non-array values will be converted to arrays. The loop node ends when the loop count is reached, or when the array loop is completed. You can also add condition nodes to the loop to terminate it.": "单一数字值将被视为循环次数，单一字符串值将被视为字符数组，其他非数组值将被转换为数组。达到循环次数，或者将数组循环完成后，循环节点结束。你也可以在循环中添加条件节点，以终止循环。"
+  "A single number will be treated as a loop count, a single string will be treated as an array of characters, and other non-array values will be converted to arrays. The loop node ends when the loop count is reached, or when the array loop is completed. You can also add condition nodes to the loop to terminate it.": "单一数字值将被视为循环次数，单一字符串值将被视为字符数组，其他非数组值将被转换为数组。达到循环次数，或者将数组循环完成后，循环节点结束。你也可以在循环中添加条件节点，以终止循环。",
+  "Start index": "起始索引",
+  "The index number used in loop scope variable.": "循环作用域变量中使用的索引数字。",
+  "From 0": "从 0 开始",
+  "From 1": "从 1 开始",
+  "Follow programming language conventions.": "遵循编程语言习惯。",
+  "Follow natural language conventions.": "遵循自然语言习惯。",
+  "Loop condition on each item": "循环每一项的条件",
+  "When to check": "检查时机",
+  "Before each starts": "每一轮开始之前",
+  "After each ends": "每一轮完成之后",
+  "When condition is not met on item": "当项不满足条件时",
+  "Exit loop": "退出循环",
+  "Continue on next item": "继续下一项",
+  "Condition": "条件",
+  "When nodes inside loop failed": "当循环内节点执行失败时",
+  "Continue loop on next item": "继续循环下一项",
+  "Exit loop and continue workflow": "退出循环并继续工作流",
+  "Exit workflow": "退出工作流"
 }
