+++ packages/plugins/@nocobase/plugin-workflow-loop/src/server/LoopInstruction.ts
@@ -7,7 +7,9 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-import { Processor, Instruction, JOB_STATUS, FlowNodeModel, JobModel } from '@nocobase/plugin-workflow';
+import evaluators from '@nocobase/evaluators';
+import { Processor, Instruction, JOB_STATUS, FlowNodeModel, JobModel, logicCalculate } from '@nocobase/plugin-workflow';
+import { EXIT } from '../constants';
 
 function getTargetLength(target) {
   let length = 0;
@@ -23,33 +25,49 @@ function getTargetLength(target) {
   return length;
 }
 
+function calculateCondition(node: FlowNodeModel, processor: Processor) {
+  const { engine, calculation, expression } = node.config || {};
+  const evaluator = evaluators.get(engine);
+
+  return evaluator
+    ? evaluator(expression, processor.getScope(node.id, true))
+    : logicCalculate(processor.getParsedValue(calculation, node.id, true));
+}
+
 export default class extends Instruction {
   async run(node: FlowNodeModel, prevJob: JobModel, processor: Processor) {
     const [branch] = processor.getBranches(node);
     const target = processor.getParsedValue(node.config.target, node.id);
     const length = getTargetLength(target);
+    const looped = 0;
 
     if (!branch || !length) {
       return {
         status: JOB_STATUS.RESOLVED,
-        result: 0,
+        result: { looped },
       };
     }
 
+    // NOTE: save job for condition calculation
     const job = await processor.saveJob({
       status: JOB_STATUS.PENDING,
-      // save loop index
-      result: 0,
+      result: { looped },
       nodeId: node.id,
       nodeKey: node.key,
       upstreamId: prevJob?.id ?? null,
     });
 
-    // TODO: add loop scope to stack
-    // processor.stack.push({
-    //   label: node.title,
-    //   value: node.id
-    // });
+    const { checkpoint, calculation, expression, continueOnFalse } = node.config;
+    if ((calculation || expression) && !checkpoint) {
+      const condition = calculateCondition(node, processor);
+      if (!condition && !continueOnFalse) {
+        job.set({
+          status: JOB_STATUS.RESOLVED,
+          result: { looped, broken: true },
+        });
+        return job;
+      }
+    }
 
     await processor.run(branch, job);
 
@@ -62,39 +80,68 @@ export default class extends Instruction {
     const [branch] = processor.getBranches(node);
 
     const { result, status } = job;
-    // if loop has been done (resolved / rejected), do not care newly executed branch jobs.
+    // NOTE: if loop has been done (resolved / rejected), do not care newly executed branch jobs.
     if (status !== JOB_STATUS.PENDING) {
       return processor.exit();
     }
 
-    const nextIndex = result + 1;
+    const nextIndex = result.looped + 1;
 
     const target = processor.getParsedValue(loop.config.target, node.id);
-    // branchJob.status === JOB_STATUS.RESOLVED means branchJob is done, try next loop or exit as resolved
-    if (branchJob.status > JOB_STATUS.PENDING) {
-      job.set({ result: nextIndex });
+    // NOTE: branchJob.status === JOB_STATUS.RESOLVED means branchJob is done, try next loop or exit as resolved
+    if (
+      branchJob.status === JOB_STATUS.RESOLVED ||
+      (branchJob.status < JOB_STATUS.PENDING && loop.config.exit === EXIT.CONTINUE)
+    ) {
+      const { calculation, expression, continueOnFalse } = loop.config;
+
+      job.set({ result: { looped: nextIndex } });
+      await processor.saveJob(job);
+
+      if (calculation || expression) {
+        const condition = calculateCondition(loop, processor);
+        if (!condition && !continueOnFalse) {
+          job.set({
+            status: JOB_STATUS.RESOLVED,
+            result: { looped: nextIndex, broken: true },
+          });
+          return job;
+        }
+      }
 
       const length = getTargetLength(target);
       if (nextIndex < length) {
         await processor.saveJob(job);
         await processor.run(branch, job);
-        return null;
+        return processor.exit();
+      } else {
+        job.set({
+          status: JOB_STATUS.RESOLVED,
+        });
       }
+    } else {
+      // NOTE: branchJob.status < JOB_STATUS.PENDING means branchJob is rejected, any rejection should cause loop rejected
+      job.set(
+        loop.config.exit
+          ? {
+              result: { looped: result.looped, broken: true },
+              status: JOB_STATUS.RESOLVED,
+            }
+          : {
+              status: branchJob.status,
+            },
+      );
     }
-
-    // branchJob.status < JOB_STATUS.PENDING means branchJob is rejected, any rejection should cause loop rejected
-    job.set({
-      status: branchJob.status,
-    });
 
     return job;
   }
 
-  getScope(node, index, processor) {
+  getScope(node, { looped }, processor) {
     const target = processor.getParsedValue(node.config.target, node.id);
     const targets = (Array.isArray(target) ? target : [target]).filter((t) => t != null);
     const length = getTargetLength(target);
-    const item = typeof target === 'number' ? index : targets[index];
+    const index = looped + (node.config.startIndex ? 1 : 0);
+    const item = typeof target === 'number' ? index : targets[looped];
 
     const result = {
       item,
