+++ packages/plugins/@nocobase/plugin-workflow-request/src/locale/zh-CN.json
@@ -12,8 +12,8 @@
   "Add key-value pairs": "添加键值对",
   "Format": "格式化",
   "Insert": "插入",
-  "Timeout config": "超时设置",
-  "ms": "毫秒",
+  "Timeout": "超时设置",
+  "Milliseconds": "毫秒",
   "Input request data": "输入请求数据",
   "Only support standard JSON data": "仅支持标准 JSON 数据",
   "\"Content-Type\" will be ignored from headers.": "请求头中配置的 \"Content-Type\" 将被忽略。",
